+++ sdk/resourcemanager/managedservices/armmanagedservices/zz_generated_marketplaceregistrationdefinitionswithoutscope_client.go
@@ -36,7 +36,7 @@ func NewMarketplaceRegistrationDefinitionsWithoutScopeClient(credential azcore.T
 	if options == nil {
 		options = &arm.ClientOptions{}
 	}
-	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	ep := cloud.AzurePublic.Services[cloud.ResourceManager].Endpoint
 	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
 		ep = c.Endpoint
 	}
@@ -53,6 +53,7 @@ func NewMarketplaceRegistrationDefinitionsWithoutScopeClient(credential azcore.T
 
 // Get - Get the marketplace registration definition for the marketplace identifier.
 // If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2022-01-01-preview
 // marketplaceIdentifier - The Azure Marketplace identifier. Expected formats: {publisher}.{product[-preview]}.{planName}.{version}
 // or {publisher}.{product[-preview]}.{planName} or {publisher}.{product[-preview]} or
 // {publisher}).
@@ -87,7 +88,7 @@ func (client *MarketplaceRegistrationDefinitionsWithoutScopeClient) getCreateReq
 	reqQP := req.Raw().URL.Query()
 	reqQP.Set("api-version", "2022-01-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header.Set("Accept", "application/json")
+	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
 }
 
@@ -102,10 +103,11 @@ func (client *MarketplaceRegistrationDefinitionsWithoutScopeClient) getHandleRes
 
 // NewListPager - Gets a list of the marketplace registration definitions for the marketplace identifier.
 // If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2022-01-01-preview
 // options - MarketplaceRegistrationDefinitionsWithoutScopeClientListOptions contains the optional parameters for the MarketplaceRegistrationDefinitionsWithoutScopeClient.List
 // method.
 func (client *MarketplaceRegistrationDefinitionsWithoutScopeClient) NewListPager(options *MarketplaceRegistrationDefinitionsWithoutScopeClientListOptions) *runtime.Pager[MarketplaceRegistrationDefinitionsWithoutScopeClientListResponse] {
-	return runtime.NewPager(runtime.PageProcessor[MarketplaceRegistrationDefinitionsWithoutScopeClientListResponse]{
+	return runtime.NewPager(runtime.PagingHandler[MarketplaceRegistrationDefinitionsWithoutScopeClientListResponse]{
 		More: func(page MarketplaceRegistrationDefinitionsWithoutScopeClientListResponse) bool {
 			return page.NextLink != nil && len(*page.NextLink) > 0
 		},
@@ -145,7 +147,7 @@ func (client *MarketplaceRegistrationDefinitionsWithoutScopeClient) listCreateRe
 	}
 	reqQP.Set("api-version", "2022-01-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header.Set("Accept", "application/json")
+	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
 }